requirements.txt
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.5
django-registration==2.4.1
django-tinymce==2.7.0